Update frontend services so they work with the recent Backend version.
